+++ api/common/string_type.ts
@@ -1,3 +1,5 @@
 export type BodyType = 'json' | 'xml' | 'text';
 
 export type ProjectFolderType = 'lib' | 'data';
+
+export type ImportType = 'swagger' | 'postman';