+++ client/src/common/body_type.ts
@@ -1,6 +1,6 @@
 export const bodyTypes = {
     'application/json': 'json',
-    'application/javascript': 'javascript',
+    'application/javascript': 'json',
     'application/xml': 'xml',
     'text/xml': 'xml',
     'text/plain': 'text',